Rewrite debian/copyright to use the new format
Add lots of additional copyright information, cutting and pasting
heavily from the similar work on xmltooling.  Add a note that the
source was repackaged to remove the WS-Trust.xsd file, which isn't
true yet but will be shortly.  Change the Debian packaging to the
Expat license (will confirm with Ferenc).