Update build-depends and re-run libtool at build time
Add unixodbc-dev to build-depends so that we build with ODBC support
as well.  Alphabetize the build-dependencies and add the rest of the
autotools.  Run libtool and the other autotools at build time to avoid
the AMD64 rpath bug.  Explicitly build with Apache 2.2.  Use the
Debian unstable package name for Xerces-C.