Move the Shibboleth SP schemas into a separate package
It looks like the library itself wants to have the schemas available, so
to be fully correct, break them into a separate package that the library
can depend on rather than putting them in with the Apache module.  Since
the Apache module is currently the only user of the library, this won't
make a difference in practice, but I think it more accurately reflects
the dependency structure.